data_mock/requirements.txt
pandas>=1.5.0
tabula-py>=2.5.0

